Create game with IOT
A game for children that includes IoT and AI models. There will be cases in which the child jumps on the carpet and chooses the best solution to each case. Among the cases are harassment, the way to manage money, and the way to deal with the elderly, and we can add more cases over time.